[CP] Introduce flex_cp_forward custom op for FlexAttention CP
The custom op will fetch the required K and V. Currently, the forward pass is just an all-gather, and the backward pass is a reduce-scatter.  While the logic is the same as all_gather_tensor_autograd, the custom op avoids the Autograd warning that wait_tensor() is registered to autograd.

For the next step, we should explore how to interpolate the required communication based on the information from BlockMask.

ghstack-source-id: a3dd785fc7277e384dd73d99f1943d39fc2c0a7d
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/163185